style_app/requirements.txt
flask
Pillow
tensorflow
tensorflow_hub
jupyter
keras
numpy
matplotlib